refactor(llm-cli): split conversation module into submodules
