+++ crates/llm-cli/src/components/mod.rs
@@ -1,5 +1,3 @@
-pub mod conversation;
 pub mod input;
 
-pub use conversation::Conversation;
 pub use input::Prompt;
