+++ crates/llm-cli/src/conversation/mod.rs
@@ -1,0 +1,14 @@
+mod assistant_block;
+mod conversation;
+mod mutate;
+mod node;
+mod thought_step;
+mod tool_step;
+mod user_bubble;
+
+pub use assistant_block::AssistantBlock;
+pub use conversation::Conversation;
+pub use node::Node;
+pub use thought_step::ThoughtStep;
+pub use tool_step::ToolStep;
+pub use user_bubble::UserBubble;
